fix: get contractDeployments
